Insert cast: long -> int

+++ MicroOrm.Dapper.Repositories/Repositories/DapperRepository.cs
@@ -80,7 +80,7 @@ namespace MicroOrm.Dapper.Repositories.Repositories
 
             if (SqlGenerator.IsIdentity)
             {
-                var newId = Connection.Query<long>(queryResult.Sql, queryResult.Param).FirstOrDefault();
+                var newId = Connection.Query<int>(queryResult.Sql, queryResult.Param).FirstOrDefault();
                 added = newId > 0;
 
                 if (added)
@@ -105,7 +105,7 @@ namespace MicroOrm.Dapper.Repositories.Repositories
 
             if (SqlGenerator.IsIdentity)
             {
-                var newId = (await Connection.QueryAsync<long>(queryResult.Sql, queryResult.Param)).FirstOrDefault();
+                var newId = (await Connection.QueryAsync<int>(queryResult.Sql, queryResult.Param)).FirstOrDefault();
                 added = newId > 0;
 
                 if (added)
